@@ -1,99 +1,81 @@
-﻿import common = require("./pulltorefresh-common");
-import dependencyObservable = require("ui/core/dependency-observable");
-import proxy = require("ui/core/proxy");
-import view = require("ui/core/view");
-import style = require("ui/styling/style");
+﻿import { PullToRefreshBase, refreshingProperty, colorProperty, backgroundColorProperty } from "./pulltorefresh-common";
+import { Color } from "color";
 
+export * from "./pulltorefresh-common";
 
-function refreshingPropertyChanged(data: dependencyObservable.PropertyChangeData) {
-    var pullRefresh = <PullToRefresh>data.object;
-    if (!pullRefresh.android) {
-        return;
-    }
-
-    pullRefresh.android.setRefreshing(data.newValue);
-}
-
-// register the setNativeValue callback
-(<proxy.PropertyMetadata>common.PullToRefresh.refreshingProperty.metadata).onSetNativeValue = refreshingPropertyChanged;
-
-
-global.moduleMerge(common, exports);
-
-export class PullToRefresh extends common.PullToRefresh {
-    private _android: android.support.v4.widget.SwipeRefreshLayout;
+export class PullToRefresh extends PullToRefreshBase {
     private _androidViewId: number;
-    
-    constructor() {
-        super();
-    }
+
+    public nativeView: android.support.v4.widget.SwipeRefreshLayout;
 
     get android(): android.support.v4.widget.SwipeRefreshLayout {
-        return this._android;
+        return this.nativeView;
     }
-
-    get _nativeView(): android.support.v4.widget.SwipeRefreshLayout {
-        return this._android;
-    }
-
-    public _createUI() { 
-
-        var that = new WeakRef(this);
-
-        this._android = new android.support.v4.widget.SwipeRefreshLayout(this._context);
+    
+    public createNativeView() { 
+        const swipeRefreshLayout = new android.support.v4.widget.SwipeRefreshLayout(this._context);
 
         if (!this._androidViewId) {
             this._androidViewId = android.view.View.generateViewId();
         }
-        this._android.setId(this._androidViewId);
+        swipeRefreshLayout.setId(this._androidViewId);
 
-        this._android.setOnRefreshListener(new android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener({
-            get owner() {
-                return that.get();
-            },
+        const refreshListener = new SewipeRefreshListener(new WeakRef(this));
+        swipeRefreshLayout.setOnRefreshListener(refreshListener);
+        (swipeRefreshLayout as any).refreshListener = refreshListener;
 
-            onRefresh: function (v) {
-                var owner = that.get();
-                if (owner) {
-                    owner.refreshing = true;
-                    owner._emit(common.PullToRefresh.refreshEvent);
-                }
-            }
-        }));
+        return swipeRefreshLayout;
+    }
+
+    public initNativeView() {
+        super.initNativeView();
+
+        const nativeView = this.nativeView as any;
+        nativeView.refreshListener.owner = new WeakRef(this);
+    }
+
+    public disposeNativeView() {
+        const nativeView = this.nativeView as any;
+        nativeView.refreshListener.owner = null;
+        
+        super.disposeNativeView();
+    }
+
+    [refreshingProperty.getDefault](): boolean {
+        return false;
+    }
+    [refreshingProperty.setNative](value: boolean) {
+        this.nativeView.setRefreshing(value);
+    }
+    
+    [colorProperty.setNative](value: Color | number) {
+        const color = value instanceof Color ? value.android : value;
+        this.nativeView.setProgressBackgroundColorSchemeColor(color);
+    }
+
+    [backgroundColorProperty.setNative](value: Color | number) {
+        const color = value instanceof Color ? value.android : value;
+        this.nativeView.setColorSchemeColors(color);
     }
 }
 
-export class PullToRefreshStyler implements style.Styler {
-    private static setBackgroundColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <android.support.v4.widget.SwipeRefreshLayout>pullToRefresh._nativeView;
-        native.setProgressBackgroundColorSchemeColor(value);
+@Interfaces([android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener])
+class SewipeRefreshListener extends java.lang.Object implements android.support.v4.widget.SwipeRefreshLayout.OnRefreshListener {
+    constructor(private owner: WeakRef<PullToRefresh>) {
+        super();
+
+        return global.__native(this);
     }
-    private static resetBackgroundColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <android.support.v4.widget.SwipeRefreshLayout>pullToRefresh._nativeView;
-        native.setProgressBackgroundColorSchemeColor(value);
-    }
-    
-    private static setColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <android.support.v4.widget.SwipeRefreshLayout>pullToRefresh._nativeView;        
-        native.setColorSchemeColors([value]);
-    }
-    private static resetColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <android.support.v4.widget.SwipeRefreshLayout>pullToRefresh._nativeView;
-        native.setColorSchemeColors([value]);
-    }
-    
-    public static registerHandlers() {
-        style.registerHandler(style.backgroundColorProperty, 
-            new style.StylePropertyChangedHandler(PullToRefreshStyler.setBackgroundColor, 
-                PullToRefreshStyler.resetBackgroundColor), 
-            "PullToRefresh");
-        style.registerHandler(style.backgroundInternalProperty, 
-            style.ignorePropertyHandler, 
-            "PullToRefresh");
-        style.registerHandler(style.colorProperty, 
-            new style.StylePropertyChangedHandler(PullToRefreshStyler.setColor, 
-                PullToRefreshStyler.resetColor), 
-            "PullToRefresh");
+
+    public onRefresh(v) {
+        var owner = this.owner.get();
+
+        if (owner) {
+            owner.refreshing = true;
+            owner.notify({
+                eventName: PullToRefreshBase.refreshEvent,
+                object: owner
+            });
+        }
     }
 }
-PullToRefreshStyler.registerHandlers();
