@@ -1,26 +1,7 @@
-import common = require("./pulltorefresh-common");
-import proxy = require("ui/core/proxy");
-import dependencyObservable = require("ui/core/dependency-observable");
-import style = require("ui/styling/style");
+import { PullToRefreshBase, refreshingProperty, colorProperty, backgroundColorProperty } from "./pulltorefresh-common";
+import { Color } from "color";
 
-function refreshingPropertyChanged(data: dependencyObservable.PropertyChangeData) {
-    var pullRefresh = <PullToRefresh>data.object;
-    if (!pullRefresh.refreshControl) {
-        return;
-    }
-
-    if (data.newValue) {
-        pullRefresh.refreshControl.beginRefreshing();
-    }
-    else {
-        pullRefresh.refreshControl.endRefreshing();
-    }
-}
-
-// register the setNativeValue callback
-(<proxy.PropertyMetadata>common.PullToRefresh.refreshingProperty.metadata).onSetNativeValue = refreshingPropertyChanged;
-
-global.moduleMerge(common, exports);
+export * from "./pulltorefresh-common";
 
 class PullToRefreshHandler extends NSObject {
     public static ObjCExposedMethods = {
@@ -38,27 +19,28 @@ class PullToRefreshHandler extends NSObject {
     public handleRefresh(refreshControl: UIRefreshControl) {
         var pullToRefresh = this._owner.get();
         pullToRefresh.refreshing = true;
-        pullToRefresh._emit(common.PullToRefresh.refreshEvent);
+        pullToRefresh.notify({
+            eventName: PullToRefreshBase.refreshEvent,
+            object: pullToRefresh
+        });
     }
 }
 
-export class PullToRefresh extends common.PullToRefresh {
-    private _refreshControl: UIRefreshControl;
+export class PullToRefresh extends PullToRefreshBase {
     private _handler: PullToRefreshHandler;
+
+    // NOTE: We ccannot use the default ios property as the UIRefreshControl can be added only to UIScrollViews!
+    public refreshControl: UIRefreshControl;
     
     constructor() {
         super();
         
-        this._refreshControl = new UIRefreshControl();
+        this.refreshControl = UIRefreshControl.alloc().init();
         this._handler = PullToRefreshHandler.initWithOnwer(new WeakRef(this));
-        this._refreshControl.addTargetActionForControlEvents(this._handler, "handleRefresh", UIControlEvents.ValueChanged);
+        this.refreshControl.addTargetActionForControlEvents(this._handler, "handleRefresh", UIControlEvents.ValueChanged);
     }
     
-    // NOTE: We ccannot use the default ios property as the UIRefreshControl can be added only to UIScrollViews!
-    get refreshControl(): UIRefreshControl {
-        return this._refreshControl
-    }
-        
+
     public onLoaded() {
         super.onLoaded();
         
@@ -66,51 +48,46 @@ export class PullToRefresh extends common.PullToRefresh {
             // Ensure that we can trigger the refresh, even if the content is not large enough
             this.content.ios.alwaysBounceVertical = true;
             
-            this.content.ios.addSubview(this._refreshControl);
+            this.content.ios.addSubview(this.refreshControl);
         }
         else if (this.content.ios instanceof UIWebView) {
             // Ensure that we can trigger the refresh, even if the content is not large enough
             this.content.ios.scrollView.alwaysBounceVertical = true;
             
-            this.content.ios.scrollView.addSubview(this._refreshControl);
+            this.content.ios.scrollView.addSubview(this.refreshControl);
         }
         else {
             throw new Error("Content must inherit from UIScrollView!");
         }
     }
-}
 
-export class PullToRefreshStyler implements style.Styler {
-    private static setBackgroundColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <UIRefreshControl>pullToRefresh.refreshControl;
-        native.backgroundColor = value;
+    [refreshingProperty.getDefault](): boolean {
+        return false;
     }
-    private static resetBackgroundColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <UIRefreshControl>pullToRefresh.refreshControl;
-        native.backgroundColor = value;
-    }
-    
-    private static setColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <UIRefreshControl>pullToRefresh.refreshControl;
-        native.tintColor = value;
-    }
-    private static resetColor(pullToRefresh: PullToRefresh, value: any) {
-        var native = <UIRefreshControl>pullToRefresh.refreshControl;
-        native.tintColor = value;
+    [refreshingProperty.setNative](value: boolean) {
+        if (value) {
+            this.refreshControl.beginRefreshing();
+        }
+        else {
+            this.refreshControl.endRefreshing();
+        }
     }
     
-    public static registerHandlers() {
-        style.registerHandler(style.backgroundColorProperty, 
-            new style.StylePropertyChangedHandler(PullToRefreshStyler.setBackgroundColor, 
-                PullToRefreshStyler.resetBackgroundColor), 
-            "PullToRefresh");
-        style.registerHandler(style.backgroundInternalProperty, 
-            style.ignorePropertyHandler, 
-            "PullToRefresh");
-        style.registerHandler(style.colorProperty, 
-            new style.StylePropertyChangedHandler(PullToRefreshStyler.setColor, 
-                PullToRefreshStyler.resetColor), 
-            "PullToRefresh");
+    [colorProperty.getDefault](): UIColor {
+        return this.refreshControl.tintColor;
     }
+    [colorProperty.setNative](value: Color | UIColor) {
+        const color = value instanceof Color ? value.ios : value;
+
+        this.refreshControl.tintColor = color;
+    }
+
+    [backgroundColorProperty.getDefault](): UIColor {
+        return this.refreshControl.backgroundColor;
+    }
+    [backgroundColorProperty.setNative](value: Color | UIColor) {
+        const color = value instanceof Color ? value.ios : value;
+
+        this.refreshControl.backgroundColor = color;
+    }    
 }
-PullToRefreshStyler.registerHandlers();
